[GraphTrainer] Rename codegen.py -> custom_codegen.py and test accordingly
